servers/pricing-strategy/pyproject.toml
[project]
name = "pricing-strategy"
version = "0.1.0"
description = "Pricing Strategy MCP Server with LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.13.1",
    "langgraph>=0.2.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]

